# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News aggregator API documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/preferences": {
            "get": {
                "tags": [
                    "preferences"
                ],
                "summary": "user preferences",
                "operationId": "df6c970adbe38ff2fd32bb3d4d02739e",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "data": {
                                                "preference_type": "author",
                                                "0": "Koelpin"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "preferences"
                ],
                "summary": "set user preferences",
                "operationId": "7eb122f4d23d7acfca2cc456a5f78d97",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "preferences": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "preference_type": {
                                                    "type": "enum",
                                                    "example": 1
                                                },
                                                "preference_value": {
                                                    "type": "string",
                                                    "example": "Koelpin"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "data": [],
                                            "message": "Your preferences have been set successfully."
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "preferences"
                ],
                "summary": "delete user preferences",
                "operationId": "432fde41e0548b47bb540ad49ebce502",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "data": [],
                                            "message": "Your preferences have been deleted successfully"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}